Feature/add edit components

+++ server/db/models/Product.js
@@ -1,8 +1,8 @@
-const Sequelize = require('sequelize');
-const { Module } = require('webpack');
-const db = require('../db');
+const Sequelize = require("sequelize");
+const { Module } = require("webpack");
+const db = require("../db");
 
-const Product = db.define('product', {
+const Product = db.define("product", {
   name: {
     type: Sequelize.STRING,
     allowNull: false,
@@ -11,10 +11,10 @@ const Product = db.define('product', {
     },
   },
   type: {
-    type: Sequelize.ENUM('milk', 'white', 'dark', 'assorted'),
+    type: Sequelize.ENUM("milk", "white", "dark", "assorted"),
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER,
     validate: {
       min: 0.0,
     },
